changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11758 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Misc Objects/Rates/ORTimeRate.h
@@ -44,9 +44,9 @@
 - (void) setLastAverageTime:(NSDate*)newLastAverageTime;
 - (unsigned long) sampleTime;
 - (void) setSampleTime:(unsigned long)newSampleTime;
-- (unsigned) count;
-- (double)valueAtIndex:(unsigned)index;
-- (NSTimeInterval)timeSampledAtIndex:(unsigned)index;
+- (NSUInteger) count;
+- (double)valueAtIndex:(NSUInteger)index;
+- (NSTimeInterval)timeSampledAtIndex:(NSUInteger)index;
 - (NSArray*) ratesAsArray;
